refactor: clean up review feedback
- remove no-op test wrapper fns in projections/regex_projections tests
- move test_handle_histogram to sql/histogram.rs alongside the function
- define DEFAULT_ORG once in config and re-export from core/vrl
