+++ src/core/src/common/meta/organization.rs
@@ -13,14 +13,13 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+pub use config::DEFAULT_ORG;
 use config::{meta::user::UserRole, stats::MemorySize};
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 
 #[cfg(feature = "cloud")]
 use super::user::InviteStatus;
-
-pub const DEFAULT_ORG: &str = "default";
 pub const CUSTOM: &str = "custom";
 pub const USER_DEFAULT: &str = "user_default";
 pub const THRESHOLD: i64 = 9383939382;
